pull user-friendly function out of the specific submodule
